use index.tsx directly, if you have use them at all
